First benchmark series -- check.
MongoDB running on a 1-core VM on the same machine:

(documents resemble a SAGA-Pilot work-unit)

Results for mongodb://mongohost:27017/:
======================================================================
Average time to create a collection: 0.001537 sec.
Average time to add documents in bulks of 1: 0.000535 sec.
Average time to add documents in bulks of 2: 0.000353 sec.
Average time to add documents in bulks of 4: 0.000501 sec.
Average time to add documents in bulks of 8: 0.000754 sec.
Average time to add documents in bulks of 16: 0.000905 sec.
Average time to add documents in bulks of 32: 0.001462 sec.
Average time to add documents in bulks of 64: 0.002336 sec.
Average time to add documents in bulks of 128: 0.004139 sec.
Average time to add documents in bulks of 256: 0.008066 sec.
Average time to add documents in bulks of 512: 0.017106 sec.
Average time to add documents in bulks of 1024: 0.034532 sec.
Average time to add documents in bulks of 2048: 0.077557 sec.
Average time to add documents in bulks of 4096: 0.147238 sec.
Average time to add documents in bulks of 8192: 2.678289 sec.
Average time to add documents in bulks of 16384: 3.671604 sec.