Added a dummy / skeleton agent + example
The skeleton agent (~350 LOC) launches on a remote site and pulls jobs from MongoDB. However, it just pretends to execute the tasks. It simply sets the task states to 'DONE' without ever laumching the executable.

The skeleton agent is a great starting point for developing a new agent for RADICAL-Pilot as it comes without all the clutter of the 'real' RADICAL_pilot agent. The command-line interface, etc. is all the same though.

The only thing that you have to do in order to use the agent is to add the following lines to your pilot description:

    pdesc.cleanup  = False
    pdesc.pilot_agent_priv = "radical-pilot-agent-skeleton.py"

The 'cleanup' directive is not strictly necessary, but it prevents the agent wrapper script to clean up the sandbox directory. That of course comes in very handy for debugging purposes. Run the example in tests/simple-skeleton.py to see what it does / how it works.