Bootstrapper and agent major revision.
- Removed most of the lrms and launch method discovery code, lrms and launch
methods are now specified in the resource config.
Corresponding command lines are still discovered. (See below for resource
config options)

- Node and core discovery is now nicely splitted over lrms's in separate methods.

- Most uses of "host" is now "node" to make it more generic where there are
multiple vnodes per host.

- Fixed an off-by-one error in the core/node acquisition.

- Remove all references to unitmanager in the agent.

- Implement local execution.

- Agent and bootstrapper debug level can now be set by the application.

- Removed passing of arguments to bootstrapper and agent respectively that were
not used.

- Ordered arguments to keep better track of them as we are approaching full
alfabet.

- Bootstrapper can now make use of a pre-populated virtual environment that is
passed as an argument, and will skip the installation part of the bootstrap in
that case.

- Allow the setup of an ssh tunnel to a configurable endpoint during the
bootstrap that will tunnel the mongodb connection and ideally also the
installation with pip and easy_install. The tunnelling of the installation is
not working as it should, mainly due to the sorry state of pip and easy_install
themselves.

- Updated some of the resource configuration files, will update the others when I
test more resources.

- Some variable renaming here and there for better match with their semantics.

Tested on Stampede, India, Yellowstone, Archer, DAS4 so far.

New Resource Configuration options:

- 'global_virtenv': Set a path to an existing pre-populated virtual environment.
This one will be sources and the installation step of the bootstrapper will be
skipped.

- 'lrms': Specify the type of the LRMS. Possible values: TORQUE, PBSPRO, SLURM,
SGE, LSF, FORK.

- 'task_launch_method': The method to launch regular tasks. Possible values:
LOCAL, SSH, APRUN.

- 'mpi_launch_method': The method to launch MPI tasks. Possible values: MPIRUN,
APRUN, IBRUN, POE.

- 'forward_tunnel_endpoint': When the host where the agent runs has no outbound
network connectivity, a tunnel can be setup so that the bootstrapper can
download and the mongodb can be reached.