redefine what a schedule is
a unit can now actually enter SCHEDULED state, which was defined but never used.  OTOH, it can also remain in NEW state if it is, at this point, *not* scheduled to a pilot.
That change required additional methods in the controller and the db layer to publish CUs which are not assigned to a pilot.