fix for #330
you can define a 'mandatory_args' key in the resource config now. All
elements that are defined as that and are not defined in the
ComputePilotDescroption raise a BadParameter description.

This has been exemplarily implemented for stampede and archer ("project"). Other
resource configurations can be amended like this on demand.