Fix off-by-one error.
This resolves #621 and #572.

My hunch is that there still is an underlying issue in the multinode scheduler,
will create a separate ticket for that.

Andre: To answer your (now removed question), a non-communicating task
can not have concurrency beyond node boundary. This seems obvious, so I might
miss the point of your question?