Really disable timeout mechanism.
Given that this was not used anyway, and causes problems on BW, there is no
reason not to do this.

Symptoms seen are that the timeout that sends the alarm causes the script to
exit. Which make me think that that my '/bin/sleep' might do something funky
with the SIGALRM handler.

Ultimately it would be good to better understand it, but this is not the time
and place.