introduce a barrier on agent startup
this barrier will wait for each agent instance until all components and workers
sent an ALIVE message on the command pubsub.  Similar messages are also sent for
each spawned sub-agent, so that ultimately the barrier will only get released
once all components for all agents are up.

The barrier does not include the communication bridges (because it *uses* them).

The above implies that all components now publish on the command channel.  They
also all have subscribed a command_cb on that channel which will listen for
terminate commands.  What is missing right now is the timely creation of
shutdown commands on component failures -- that should go into
Component._finalize().