3 days of debugging, three lines of code removed.  Seems fair.
I don't think I earned the Haribo though, not yet at least, as I don't
understand yet what really happens.  Those three lines should have led to
*additional*, not missing alive messages.  Well, the problem was not really the
alive message, but that we create a ZMQ pubsub subscription on a channel, and
pass it opened to a child (via fork), and then in the child re-create the same
channel.  ZMQ does not seem to like this, and got hung up on some initialization
routine or something, so that we never got the child pubsub up, and *that* alive
message was then missing.