bridges now find free ports on their own
This implies that we need to inspect the bridges for the used addresses
afterwards, and that we need to communicate both ends (sink and source) to the
agent components individually.

Further, as the addresses are now only known *after* starting the bridges, we
need to move the startup into bootstrap_3, before creating the AgentWorker, as
that already tries to connect to the bridges.

FIXME: bridges need to be closed on termination of bootstrap_3