Add a barrier to agent startup.
This allows the application to signal the bootstrapper by means of creating a
file that it should wait with launching the agent.
Main use case is experiments, where we want all of the CUs to have reached the
database before we launch the agent, in case we want to stress test the agent.