Only break out when we can't schedule.
This is a data supported incrmental change to the recent scheduler change.
In the previous incarnation we would gradually reach a lower number of
concurrent cores as we would never allocate more than one new, even if multiple
units had completed.

This is probably not a great change though in the light of mixed size units.