fix the 'orte-dvm not in path' problem from the anaconda merge
This keeps the CU env cleanup, but tries to make sure that the sub-agents find
the same env as agent-0, by applying pre_bootstrap_1 commands also to
pre_bootstrap_2.  The assumption here is that those commands are available on
the compute nodes.

This also has some performance implications I assume -- but for a limited number
of sub-agents it should not matter,

An alternative approach would be to copy the relevant parts of the env after
pre_bootstrap_1, and applying it in bootstrap_2 before pre_bootstrap_2 commands
are executed.