Fix/macos (#1356)
* When using more than one sub-agent, pyzmq goes awol.
We avoid that in the default configuration now.  I did not manage to track this
down, ultimately.
This patch also avoids some logger re-initialization in process children,
slightly reducing the pressure on the open-file limits.  Nevertheless, we now up
that limit in the bootstrapper for Darwin targets.

Further, this patch cleans up some log initialization, to ensure that child
processes quickly switch to their own log target.

* improve a log message

* attempt to up system limits on darwin