use a separate VE for each resource config (#1367)
This removes the 'virtenv' config entry for most resource configs.  That
is now only really needed for shared, globe VEs, as on archer and
supermuc.  The new default is to use
%(global_sandbox)/ve.%(resource).%(rp_version)
as name for the VE.  This fixes two issues:

  - different resource configs need different requirements.  For example
    (and prominently), orte-cffi needs only to be installed for orte and
    ortelib configs
  - when new dependencies are introduced in a new release, then they
    will automatically be installed in the new VE

The first item is not yet taken adavntage of - for that, we need to move
the list of optional pre-requisites to the resource configs, or need to
be more clever in the bootstrapper to figure out what is actually
needed.  But the second item will apply immediately, as otherwise all
existing VEs would break on updates to 0.46.  This fixes #1363