add `deactivate` support for tasks
This patch lets the bootstrapper create a `deactivate` script during
startup, which essentially mimics the venv `deactivate` command.  Alas,
the latter is not available in forked processes, so the executor cannot
deactivate the venv cleanly for the tasks.  With this script in place,
a task can add

    `. $RP_PILOT_SANDBOX/deactivate`

to the list of `pre_exec` commands to restore the original values for
`PATH`, `PYTHONPATH` and `PYTHONHOME`.  Any following `pre_exec` which
loads a venv or conda env should then not be affected by the pilot's
venv.

We should consider adding any `module load` triggered env changes to
that deactivate script.  The difficulty here is to see what env
variables have been changed and thus need restoring.  For now, we also
log the post-bootstrap env in `env.bs_0` so that we can compare
manually.

This patch also contains some other cleanups which got carried over by
cherry-picking the deactivate patch from an unrelated branch.