move staging ops into separate worker
Oh boy, this went larger than expacted - but the rationale is this:
both pmgr and pmgr do staging on the pilot's behalf.  Further, the pilot
staged several individual files via different code paths.  Further, the
pmgr launcher had to open an addional job service to untar the pilot
sandbox.

This PR

  - moves all staging ops into a separate worker
  - moves the untar op from the pmmgr launcher to the bootstrapper

The goal is to (a) consolidate different code paths (migrating toward
a generic data stager), and (b) reduce the number of connections needed
(hi Frontera).