Encapsulate kwargs handling for Session.close().
Allow kwargs for Session.close() to be provided when initializing the Session.
This allows a "FIXME" to be resolved. When used as a Python context manager,
the Session will have user input with which to override the default arguments
at close().

If key word arguments are provided both at instantiation and when calling
`close()`, the kwargs provided to `close()` take precedence.

An attempt is made to merge the new arguments with the previous arguments.
`close()` is called with `cleanup=True` but an unspecified *terminate*, and
`cleanup=False, terminate=False` was provided to `Session()`, a warning is
issued and the old value is overwritten with `terminate=True`.

Close options are evaluated for self-consistency during both
`Session.__init__()` and `Session.close()`

## Behavior changes

If the user provides inconsistent values for *cleanup* and *terminate*,
ValueError is raised.

Previously, if the user explicitly specified `cleanup=None`, this was
interpreted as `cleanup=True` instead of the default `cleanup=False`.

Fixes #2394