Un-hide "Utilities and helpers" section in API reference.
The "Utilities and helpers" section of `apidoc.rst` was nested in the
"Tasks and TaskManagers" section (presumably by accident), and so did not
appear in the left side-bar with other major categories of API
documentation. This change elevates the heading one level to be equal with
the other major sections.